refactor(api): neutralize billing and chat metric service filenames
move three api service modules to domain-neutral filenames as git renames
and retarget the import specifiers that referenced the old paths. every
export in all three files was already neutral, so no declaration is
renamed and no billing, subscription, or metric behavior changes.

Closes #28204
