+++ turbo/apps/api/src/signals/services/chat-first-assistant-event-metric.service.ts
