+++ turbo/apps/api/src/signals/services/cron-billing-entitlements.service.ts
@@ -45,7 +45,7 @@ import {
   knownBillingPlanPriceItem,
   knownPlanPriceItem,
   tierForKnownPlanPrice,
-} from "./zero-billing-checkout.service";
+} from "./billing-checkout.service";
 import {
   reconcileUsagePackSubscriptions,
   stripeSubscriptionUsesMemberUsagePacks,
