+++ turbo/apps/api/src/signals/services/usage-pack-invitation-purchase.service.ts
@@ -55,7 +55,7 @@ import {
   isCurrentStripePreviewMetadata,
   stripePreviewMetadata,
 } from "./stripe-preview-metadata.service";
-import { activeUsagePackPriceId } from "./zero-billing-checkout.service";
+import { activeUsagePackPriceId } from "./billing-checkout.service";
 import {
   resolveBillingPurchaseRoute,
   stripeBillingPurchasePaymentParams,
