+++ turbo/apps/api/src/signals/services/usage-pack-subscription-migration.service.ts
@@ -56,7 +56,7 @@ import {
   tierForKnownPriceId,
   usagePackUsdForKnownPriceId,
   type SubscriptionCheckoutTier,
-} from "./zero-billing-checkout.service";
+} from "./billing-checkout.service";
 
 const PREVIEW_TTL_MS = 30 * 60 * 1000;
 const RECONCILIATION_DELAY_MS = 5 * 60 * 1000;
